Fix the modification of unused_disks['disk_array']
Python handle the yaml references by using itself references.
It does not copy the data inside each object. So when we
modifie the bootcmd by appending it, we will append it
for all machines that have a reference to the
unused_disks['disk_array']. This fixed it by creating a
new instance of list.