Print full error when configuration IP generates error
When we configure an IP, we can sometimes make a typo,
or take an IP of the wrong subnet. When this appens,
The full stacktrace do not show the complete message.
This is because the error returned is too long. It has the
POST url with it. This commit changes that, and before exiting,
we show only the content of the error.